Make heatmap reusable Update data intake ui
